+++ Modules/Microsoft365DSC/DSCResources/MSFT_TeamsGuestMessagingConfiguration/MSFT_TeamsGuestMessagingConfiguration.psm1
@@ -64,7 +64,11 @@ function Get-TargetResource
 
         [Parameter()]
         [Switch]
-        $ManagedIdentity
+        $ManagedIdentity,
+
+        [Parameter()]
+        [System.String[]]
+        $AccessTokens
     )
 
     Write-Verbose -Message 'Getting configuration of Teams Guest Messaging settings'
@@ -108,6 +112,7 @@ function Get-TargetResource
             TenantId               = $TenantId
             CertificateThumbprint  = $CertificateThumbprint
             ManagedIdentity        = $ManagedIdentity.IsPresent
+            AccessTokens           = $AccessTokens
         }
     }
     catch
@@ -187,7 +192,11 @@ function Set-TargetResource
 
         [Parameter()]
         [Switch]
-        $ManagedIdentity
+        $ManagedIdentity,
+
+        [Parameter()]
+        [System.String[]]
+        $AccessTokens
     )
 
     Write-Verbose -Message 'Setting configuration of Teams Guest Messaging settings'
@@ -201,6 +210,7 @@ function Set-TargetResource
     $inputValues.Remove('CertificateThumbprint') | Out-Null
     $inputValues.Remove('Identity') | Out-Null
     $inputValues.Remove('ManagedIdentity') | Out-Null
+    $inputValues.Remove('AccessTokens') | Out-Null
     foreach ($item in $inputValues)
     {
         if ([System.String]::IsNullOrEmpty($item.Value))
@@ -299,7 +309,11 @@ function Test-TargetResource
 
         [Parameter()]
         [Switch]
-        $ManagedIdentity
+        $ManagedIdentity,
+
+        [Parameter()]
+        [System.String[]]
+        $AccessTokens
     )
     #Ensure the proper dependencies are installed in the current environment.
     Confirm-M365DSCDependencies
@@ -356,7 +370,11 @@ function Export-TargetResource
 
         [Parameter()]
         [Switch]
-        $ManagedIdentity
+        $ManagedIdentity,
+
+        [Parameter()]
+        [System.String[]]
+        $AccessTokens
     )
 
     $ConnectionMode = New-M365DSCConnection -Workload 'MicrosoftTeams' `
@@ -384,6 +402,7 @@ function Export-TargetResource
             TenantId              = $TenantId
             CertificateThumbprint = $CertificateThumbprint
             ManagedIdentity       = $ManagedIdentity.IsPresent
+            AccessTokens          = $AccessTokens
         }
         $Results = Get-TargetResource @Params